+++ src/inet/networklayer/ipv6/Ipv6Header.h
@@ -69,21 +69,21 @@ class INET_API Ipv6Header : public Ipv6Header_Base
     virtual void setExplicitCongestionNotification(int ecn) override { setTrafficClass((getTrafficClass() & 0x3f) | ((ecn & 0x3) << 6)); }
 
     /** Generated but unused method, should not be called. */
-    virtual void setExtensionHeaderArraySize(unsigned int size) override;
+    virtual void setExtensionHeaderArraySize(size_t size) override;
 
     /** Generated but unused method, should not be called. */
-    virtual void setExtensionHeader(unsigned int k, Ipv6ExtensionHeader *extensionHeader_var) override;
+    virtual void setExtensionHeader(size_t k, Ipv6ExtensionHeader *extensionHeader_var) override;
 
     /**
      * Returns the number of extension headers in this datagram
      */
-    virtual unsigned int getExtensionHeaderArraySize() const override;
+    virtual size_t getExtensionHeaderArraySize() const override;
 
     /**
      * Returns the kth extension header in this datagram
      */
-    virtual Ipv6ExtensionHeader *getMutableExtensionHeader(unsigned int k) override;
-    virtual const Ipv6ExtensionHeader *getExtensionHeader(unsigned int k) const override;
+    virtual Ipv6ExtensionHeader *getMutableExtensionHeader(size_t k) override;
+    virtual const Ipv6ExtensionHeader *getExtensionHeader(size_t k) const override;
 
     /**
      * Returns the extension header of the specified type,
